fix: close stuck tooltip via a zustand store
A tooltip bound to a chat element stayed open when that element was
removed from the DOM while hovered, since mouseleave never fires for a
deleted node. Detect the detached reference in floating-ui's autoUpdate
loop and close through the same path as mouseleave.

Drive tooltip state from a module-level zustand store that lives outside
React, so it survives the controller being removed from and re-added to
the shadow DOM without leaking references into a mounted component. The
importable openTooltip/closeTooltip mutate the store directly, and the
controller wipes any open tooltip on instantiation so a remount cannot
leave one stuck open. A single TooltipController derives what it shows
from the store.
